hubble: move FakePeerNotifyServer to testutils
[ upstream commit 2064bdccfde3653e0c8f4e7f5c937cddf02563c6 ]
